move AvatarUploader and UpdatePortfolioPw to PopUp file
