+++ src/pages/EditPortfolioPage/EditPortfolioPage.js
@@ -4,8 +4,8 @@ import Container from '../../components/Container'
 import { SmallButton, LargeButton } from '../../components/buttons'
 import { SubTitle } from '../../components/heading'
 import { MEDIA_QUERY_SM } from '../../styles/breakpoints'
-import UpdatePortfolioPw from '../../components/UpdatePortfolioPw'
-import AvatarUploader from '../../components/AvatarUploader'
+import UpdatePortfolioPw from '../../components/PopUp/UpdatePortfolioPw'
+import AvatarUploader from '../../components/PopUp/AvatarUploader'
 import { Link } from 'react-router-dom'
 import { useHistory } from 'react-router'
 import { Formik, Form } from 'formik'
